Issue warning only on the first time
Temporarily have the registeredToConsume* functions only return false the first time an unregistered item is checked. This avoids having the checker issue a warning message each time. This is only temporary since this is not thread safe and in the threaded framework such a case would cause a fatal exception.