Switch from using a mutex to using a std::atomic when signaling shutdown request
To be properly thread safe under C++11 we switched from using a volatile bool to a volatile std::atomic<bool> as the way to communicate an external shutdown request. The mutex usr2_lock was also removed since it actually was never used correctly anyway and the std::atomic does the work the mutex should have been used for.
The function getSigNum() and its related global variables were removed since the function is no longer used anywhere.