Use new ThreadSafeLogMessageLoggerScribe as the default scribe
The ThreadSafeLogMessageLoggerScribe can properly handle multiple threads issuing a 'log' message to the MessageLogger. This does not require a separate thread to handle and instead one of the concurrent threads calling 'log' is picked to issue the messages for all the threads. This was made the default scribe for cmsRun.