Do not fast copy ROOT files if using multiple threads
Disabled fast copying of ROOT files when we have multiple threads and multiple Streams. This must be done since fast copying requires the output file to record all the events from the input in the same order as stored in the input. This is not guaranteed when we do multiple threads.