Fail at run time if Service uses an obsolete signal.
The threaded framework requires Services to switch to a new collection
of signals. To allow easier migration, the obsolete signals have been
replaced with a class which causes an Exception to be thrown if their
connect method is called.
Had to change one unit test which was making use of the old signals.