Handle classes use a HandleExceptionFactory
Handle classes need to be able to throw an exception in the case where the data requested was not available. The previous implementation which just held a pointer to the cms::Exception was shown to be slow in the case where users were always checking edm::Handle<>::isValid() and not dereferencing the Handle. Using a factory to construct the exception only when it is absolutely needed speeds up the code.