Multithreaded Random Number Service - Generators
First part of modifications to make the generators
use the new multithreaded random number service.
This includes all changes for the Pythia8GeneratorFilter.

HadronizerFilter and GeneratorFilter now get the random
number engine in their filter and beginLuminosityBlock
method. These engines are percolated in various ways
to where they are used. Classes are modified to only
generate random numbers in the event and beginLuminosityBlock
methods.

For Pythia8GeneratorFilter this required changes to
Pythia8Hadronizer and also the EvtGen, Photos,
and Tauola external decayers. Other generators
(for example Pythia6) are not broken by these changes,
but have not been modified yet.

HadronizerFilter and GeneratorFilter also declare
shared resources.