Make consumes call check if edm::InputTag is empty
Previously, passing an empty edm::InputTag to consumes would make the framework think you did a consumesMany (since internally an empty tag is used for that purpose). This change makes an empty tag passed to consumes behave the same way as if you had a tag for a non-existent label.