Removed FileInPath::isLocal to avoid user confusion
Several users thought FileInPath::isLocal would tell them if the file was found at all rather than just stating that the last time the file was found (say in a different job) it was found in the user’s work area. Given this was only used in a couple of places we decided it was better to just remove the method all together and require users to call FileInPath::location() if they wanted to get the same information.