Make getEngine() private in random service
In the random number generator service, make the
getEngine function which takes no arguments be
private instead of public. This function is replaced
by new getEngine functions that take either a StreamID
or a LuminosityBlockIndex as an argument. This comes
after a long migration of all clients to the new
interface. The new interace is designed to work with
the multithreaded Framework.

Note this may cause some problems when automatically
merged from 7_1_X to 7_1_THREADED_X. There will be
a separate pull request to the 7_1_THREADED_X branch
soon after the automatic merge occurs which will
resolve the problems and also delete the function
entirely on that branch.