Add support for noticing and complaining about it removed plugins.
This introduces the concept of a poisoned Plugin Cache File which will
contain dummy (wrong) entries for all plugins which should be build built
given a local area. If those plugins are actually there, nothing is
changed since the local cache file has precedence. If the plugin was not
built, e.g. it was removed in the local area, the plugin manager will
complain about not finding the dummy entry.

This patch alone will not introduce any change in behavior and will
require scram to generate the poison file in order to turn on the new
functionality.