Switch to using lower bits of cachePtr_ to identify state of RefCore cache
The cachePtr_ is used to hold either an EDProductGetter or the results of getting data from the getter. Previously the state was encoded in one of the bits in ProcessIndex. However, that structure makes it difficult to change RefCore to be thread safe. Consolidating the address and the state into one memory address makes it possible to use an std::atomic.
This design makes the assumption that the RefCore must be aligned on at least an even byte boundary. This is definitely true for all the systems we use since the 64 bit pointer requires that the class be aligned on a 64 bit boundary.