Improve exception message and comments only
This is motivated by a user question from the hypernews.
We changed the AssociationMap interface in 7_5_X in a
way that requires a change to user code. The exception
message given when the user has not changed their code
needed improvement. I also improved the comments. The
changes in this PR should not change the behavior of
cmsRun unless an exception is thrown.