Implement a well-behaved stacktrace printer.
This implements a stack trace printer which is safe to call
from a signal handler (ROOT's stack trace printer is not async-signal
safe; see 'man 7 signal' for more information about async-signal
safety).

Further, on Linux, this utilizes the clone library call directly;
clone has the advantage of avoiding the pthread_atfork handlers.
We have observed that jemalloc currently registers AS-unsafe atfork
handlers, which would cause a deadlock if we use fork().