Added edm::propagate_const
propagate_const is an C++17 experimental class used to make pointers and pointer like classes have 'deep const' semantics. This is very useful for reasoning about const thread-safety.
This is a partial implementation of the proposed standard. We will replace this class once the actual one is in the C++ standard.