Removed Transient from ProductProvenance
Transient was primarily used to do a cache of edm::Parentage even though the information is stored in ParentageRegistry. The Parentage object is only requested on output and only once per branch so the cache doesn't look to be all that useful. The cache is sometimes updated lazily which causes thread-safety issues. The easiest solution is just to remove the use of the Transient all together.
(cherry picked from commit 20c2ea942e052104142ddf96cc639e4e9d2f6319)