Make JobReport::inputFileOpened sufficiently thread-safe
Now that the mixing module runs multiple threads, we can get multiple threads each calling JobReport::inputFileOpened and trying to open files for the mixing. The code path which is called must be thread safe. Although the container used for this purpose was already thread safe, the way the item in the container was used was not safe.
(cherry picked from commit 7e85bff5864f4312a34d9ec9d65e76d3d3118f23)