Remove extraneous edges from dependency graph
Because of the way the graph is traversed, if a vertex which is part of the cycle has other dependencies (usually consumes) then those extra edges which are not part of the cycle might be included in the stack. The extra edges can make it difficult to reason about the cycle and make the output from the exception very difficult to interpret.