Regularize the handling of the first event in a lumi
The EventProcessor has to treat the first event in a lumi in a
special way because it can not ask the source what the next transition
is again since it would advance to the next transition. Previously
the code to handle this was broken into two different areas and
required the event to be handled by stream 0. This change makes the
handling of the first event closer to how other events are handled.

This required updating one unit test comparison file because the
module processing order is changed in a non-important way.