Fix module summary information
The summary information about module running is not correct.
The per path accounting was double counting each failure. This
problem was obvious.
The per module accounting was sometimes not adding up to
match the per path information (it was always less). No clear
reason was found for this, although the use of memory_order_relaxed
may be the cause (so was changed).