Allow cmsRun to have multiple LuminosityBlockPrincipals
Restructured the system to allow multiple LuminosityBlockPrincipals to be used by the EventProcessor. This required moving state information about the transition out of EventProcessor and instead pass the state info as arguments to the member functions.
In addition, the LuminosityBlockPrincipals are reused for different actual LuminosityBlock transitions. This is accomplished via the use of a ReusableObjectHolder.