Fixed LumiSummary test modules
The LumiSummary test modules had an assumption that all streams stop processing events before any stream does an endLumi call. This is not the case anymore and should not have been in the code to begin with (since it was violating the threading design).
By clearing and getting the value from the luminosity block cache atomically, we avoid a problem where the endLumi could be called while another stream is calling the event method of the module.