Set an exception on the Path early
If an exception happens on a Path, we now preset the exception to the WaitingTaskList early. This avoids problems when another Path may depend on a module which will  not run on the first Path. This can cause a ProductNotFound error on the second Path and that exception may be reported from the framework instead of the original exception which caused the problem.